Finished config reading
An error is now raised if a required parameter ('router-id', 'input-ports', 'outputs') is not given in the config file.
If an unknown parameter is given, a message is printed and the line is omitted.
parameters can be given in any order in the config file.
The first occurrence of a parameter is applied to the demon, any duplicates will be omitted.
Standard python in-line comments (e.g. # comment here) are allowed in config files, as is white space (\n characters).
Error checking is done on output ports and router ids.